Move turn taking into Entity State
To get the next turn generic over all states, there's
Entity::get_next_turn() returning Option<u32>.

This leads to a lot of unwrapping which does not make me very happy.
Maybe in action checking, the acting entity needs to be checked for
being OK (or something). Then the time can be stored/parsed in the
command.