Improve Cached Pathfinding Lookups
This is a 5x speedup of average times (1ms to
200us), measured with a Very Scientific test of
running to the bottom left, to the empty spot in
the right, then up top to gather everyone for a
big jamboree, then pathing back to the right.

But its a turn based game sooo, idk?

The cache does help. Intentionally clearing the
cache causes an average 200us slowdown.
(So, like 2x slowdown? This is "valid"-ish since
what would've been the cache would've been dropped
after every run, if the time to clear/drop the
maps matters.)